pyproject.toml
[project]
name = "mlops_system_vibe"
version = "0.1.0"
description = "build a multi-agent system that bridges generative deep learning with rigorous physical laws."
authors = [{name = "Arif"}]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "google-genai",
    "pydantic",
    "python-dotenv",
    "pyyaml",
    "uv",
    "google-adk",
    "google-genai",
    # Backend Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.7",
    # Utils
    "python-dotenv",
    "gunicorn", # For production deployment
    "httpx", # For async requests # Better logging
    "loguru",
    "torch>=2.10.0",
    "numpy>=2.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/agents", "src/core", "src/environment", "src/utils"]
